broke apicaller worker out into its own file; still tested together with its manifold
